Deleted User Management and Update Bookazon
